[update]エントランス苦戦中

+++ src/World/Node.py
@@ -1,5 +1,5 @@
 class Node:
-    def __init__(self, id: int, x: int, y: int):
+    def __init__(self, id: int, x: float, y: float):
         self.id = id
         self.x = x
         self.y = y
